Fix for newer curl (-k option)

+++ mongo.backup.sh
@@ -94,16 +94,16 @@ if [ -n "$mongodbpertableconf" ] ; then
         rm -rf "${MBD}/${db}.${coll}.${bktype}.${archname}" 2>>${rpath}/m_backup.error
         ;;
       periodic)
-        install -d "${rpath}/var/mongodb"
+        [ -d "${rpath}/var/mongodb" ] || install -d "${rpath}/var/mongodb"
         if [ -f "${rpath}/var/mongodb/${table}.lastid" ] ; then
           lastid=`cat "${rpath}/var/mongodb/${table}.lastid"`
           if [ -n "$lastid" ] ; then
             echo "no op yet"
           else
-            echo "File ${rpath}/var/mongodb/${table}.lastid exists but empty. Collection $table is not backuped!" >> ${rpath}/m_backup.error
+            echo "File ${rpath}/var/mongodb/${table}.lastid exists but empty. Collection $table is not backuped. Put 0 there to make the first-time whole table backup (echo 0 > ${rpath}/var/mongodb/${table}.lastid)" | tee -a ${rpath}/m_backup.error
           fi
         else
-          echo "File ${rpath}/var/mongodb/${table}.lastid doesn't exist. Collection $table is not backuped!" >> ${rpath}/m_backup.error
+          echo "File ${rpath}/var/mongodb/${table}.lastid doesn't exist. Collection $table is not backuped. Put 0 there to make the first-time whole table backup (echo 0 > ${rpath}/var/mongodb/${table}.lastid)" | tee -a ${rpath}/m_backup.error
         fi
         ;;
       *)
@@ -128,19 +128,22 @@ else
     fi
     
     if [ "$skipdb" == "-1" ]; then
-  ### Works if version >= 1.7, avoids intermediate space usage by uncompressed dumps
-  ### Note that it doesn't dump indexes
-  #    if [ -n "$compress" ] ; then
-  #    for collection in `mongo $DBHOST/$db --eval "db.getCollectionNames()" | tail -1 | sed 's|,| |g'` ; do
-  #      $MONGODUMP $USER $PASS --host $mongohost --db $db --collection $collection --out - | $compress > "${MBD}/${db}.${archname}/${collection}.bson.${ext}" 2>>${rpath}/m_backup.error
-  #    done
-  #    fi
-  # --------------------
-  # 
-      $MONGODUMP --host $mongohost --db $db $USER $PASS --out "${MBD}/${db}.${archname}" 1>>"$stdinto" 2>${rpath}/logs/mongo.backup.tmp && echo "mongo: $db dumped successfully" >>${rpath}/m_backup.log || echo "mongo: $db dump failed" >>${rpath}/m_backup.log
-      [ -n "$TAR" ] && $TAR "${MBD}/${db}.${archname}.tar.${ext}" "${MBD}/${db}.${archname}" 1>>"$stdinto" 2>>${rpath}/logs/mongo.backup.tmp
-      cat ${rpath}/logs/mongo.backup.tmp | grep -v ^connected | grep -v 'Removing leading' >>${rpath}/m_backup.error
-      rm -rf "${MBD}/${db}.${archname}" 2>>${rpath}/m_backup.error
+      if $compress_onthefly ; then
+### Works if version >= 1.7, avoids intermediate space usage by uncompressed dumps
+### Note that it doesn't dump indexes
+        if [ -n "$compress" ] ; then
+          for collection in `$MONGO $DBHOST/$db --quiet --eval "db.getCollectionNames()" | tail -1 | sed 's|,| |g'` ; do
+            $MONGODUMP $USER $PASS --host $mongohost --db $db --collection $collection --out - | $compress > "${MBD}/${db}.${archname}/${collection}.bson.${ext}" 2>>${rpath}/m_backup.error
+          done
+        fi
+    # --------------------
+    # 
+      else
+        $MONGODUMP --host $mongohost --db $db $USER $PASS --out "${MBD}/${db}.${archname}" 1>>"$stdinto" 2>${rpath}/logs/mongo.backup.tmp && echo "mongo: $db dumped successfully" >>${rpath}/m_backup.log || echo "mongo: $db dump failed" >>${rpath}/m_backup.log
+        [ -n "$TAR" ] && $TAR "${MBD}/${db}.${archname}.tar.${ext}" "${MBD}/${db}.${archname}" 1>>"$stdinto" 2>>${rpath}/logs/mongo.backup.tmp
+        cat ${rpath}/logs/mongo.backup.tmp | grep -v ^connected | grep -v 'Removing leading' >>${rpath}/m_backup.error
+        rm -rf "${MBD}/${db}.${archname}" 2>>${rpath}/m_backup.error
+      fi
     fi
   done
 fi
